setting:> move to common/response #2
