+++ zero-zone/src/main/java/com/dalgona/zerozone/common/response/ErrorEntity.java
@@ -1,4 +1,4 @@
-package com.dalgona.zerozone.common;
+package com.dalgona.zerozone.common.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
